requirements.txt
# Data Analysis
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Statistical Analysis
statsmodels==0.14.0
scikit-learn==1.3.2

# NLP (for complaint analysis)
nltk==3.8.1
textblob==0.17.1

# API & Web
requests==2.31.0
anthropic==0.8.1
praw==7.7.1  # Python Reddit API Wrapper

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
nbconvert==7.12.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
